Refactor question layout scaling into dedicated hook
